chore(ruff): bump 0.15.17 -> 0.16.5 and take the mechanical fixes
Bumps ruff in .pre-commit-config.yaml and in the `make lint-tools` pin, so the
local gate and the CI hooks stay in lockstep.

611 findings across 158 files are fixed here. 509 had safe fixes. The other 102
came from --unsafe-fixes, applied ONLY for rules where the transformation is
mechanical and provably behaviour-preserving:

  RUF046  redundant int() around round() / math.ceil(), both of which already
          return int on Python 3
  RUF059  unused unpacked bindings
  PIE810, C408, C409, FLY002, ISC004, SIM*, PLC0206, DTZ011, RUF007, RUF022

black and isort were re-run afterwards so the three formatters agree;
`pre-commit run --all-files` reaches a fixed point (verified: two consecutive
runs leave an identical tree).

34 findings are DEFERRED, not fixed, each with a rationale in the
[tool.ruff.lint] ignore block of the root, ai/ and tools/vmaf-tune/ configs --
ruff resolves config per-file from the nearest pyproject.toml, so all three
needed the entry. The intent is to keep this gate exactly as strict as it was
BEFORE the bump: a version upgrade should not smuggle in a refactor. None of
the 34 flags a defect we know of. Two worth calling out:

  PLR0917  too-many-positional-arguments -- the positional-only variant of
           PLR0913, which this config already ignores with the note that "VMAF
           feature kernels naturally take many". Same fact, same treatment.
  BLE001   blind-except -- the batch drivers under ai/scripts/ catch Exception
           per item deliberately, so one bad row does not kill a multi-hour
           run, and predictor.py does the same to fall back to the analytical
           curve when onnxruntime is absent, mirroring the Python original.

One dead end worth recording: mid-way through, `pre-commit run --all-files`
appeared not to converge, with isort and ruff seemingly rewriting each other. I
removed "I" from ruff's select on the theory that two tools were fighting over
import order. That was wrong -- isort reports "Fixing <file>" while writing
byte-identical content, and pre-commit flags the mtime. git saw no change at
all. The select entry is restored.

+++ tools/vmaf-tune/src/vmaftune/filter_adapters/pelorus_deband.py
@@ -101,7 +101,7 @@ def _format_value(knob: Knob, value: float) -> str:
     parser) without trailing-zero noise.
     """
     if knob.is_integral:
-        return str(int(round(value)))
+        return str(round(value))
     # ``%g`` drops trailing zeros and avoids scientific notation for the
     # contract's small magnitudes (0..0.4). The contract is bit-depth
     # independent, so no locale-sensitive formatting is involved.
@@ -195,7 +195,7 @@ class PelorusDebandAdapter:
             fval = float(value)
             fval = max(knob.lo, min(knob.hi, fval))
             if knob.is_integral:
-                fval = float(int(round(fval)))
+                fval = float(round(fval))
             out[key] = fval
         return out
